Optional. Collapse InfinispanNotificationsManager into InfinispanClusterProvider
