+++ testsuite/utils/src/main/resources/log4j.properties
@@ -101,8 +101,6 @@ log4j.logger.org.apache.directory.server.ldap.LdapProtocolHandler=error
 #log4j.logger.org.keycloak.services.resources.IdentityBrokerService=trace
 #log4j.logger.org.keycloak.broker=trace
 
-#log4j.logger.org.keycloak.cluster.infinispan.InfinispanNotificationsManager=trace
-
 #log4j.logger.io.undertow=trace
 
 #log4j.logger.org.keycloak.protocol=debug
